Refactor: Move ParallelProcessor to ml_common.utils
