+++ src/utils/ml_common/validation/cv_utils.py
@@ -181,7 +181,7 @@ def _to_jsonable(obj):  # pylint: disable=too-many-return-statements
 from ..math_validation import safe_divide, safe_log  # noqa: F401 relative one-level up is correct (utils)
 from ...common_operations import create_fallback_logger
 from src.utils.hardware.m1_gpu_utils import M1GPUManager
-from src.utils.parallel_processing_optimizer import ParallelProcessor
+from src.utils.ml_common.utils import ParallelProcessor
 
 logger = _LOGGER
